requirements.txt
boto3==1.16.43
Pillow
requests